+++ Code/Subasta.Domain/Game/IFigure.cs
@@ -5,10 +5,10 @@ namespace Subasta.Domain.Game
 	public interface IFigure
 	{
 		SayKind  Say { get; }
-		int PointsBet { get; }
+		byte PointsBet { get; }
 		ICard[] MarkedCards { get; }
 		bool UsingAlternative { get; }
-		bool IsAvailable(ISaysStatus saysStatus, int normalizedTopPoints);
+		bool IsAvailable(ISaysStatus saysStatus, byte normalizedTopPoints);
 		void MarkFigures(ISaysStatus saysStatus);
 		void UnMarkPotentialCandidates();
 	}
